+++ scripts/settings.js
@@ -1,6 +1,6 @@
 async function initialize() {
-	document.querySelector('.dashboard').addEventListener('keyup', e => {if (e.which === 13) openExtensionTab('/html/dashboard.html')})
-	document.querySelector('.dashboard').addEventListener('click', _ => openExtensionTab('/html/dashboard.html'));
+	document.querySelector('.nap-room').addEventListener('keyup', e => {if (e.which === 13) openExtensionTab('/html/nap-room.html')})
+	document.querySelector('.nap-room').addEventListener('click', _ => openExtensionTab('/html/nap-room.html'));
 	showIconOnScroll();
 
 	if (window.location.hash) {
@@ -161,7 +161,7 @@ async function updateKeyBindings() {
 	commands.forEach(c => {
 		var keys = wrapInDiv('', ...splitShortcut(c.shortcut).map(s => Object.assign(document.createElement('kbd'),{innerText: s})));
 		if (choices[c.name]) bindings.append(wrapInDiv('flex', wrapInDiv({innerText: choices[c.name].label}), keys));
-		if (c.name === 'dashboard') bindings.append(wrapInDiv('flex', wrapInDiv({innerText: 'Open Sleeping Tabs'}), keys));
+		if (c.name === 'nap-room') bindings.append(wrapInDiv('flex', wrapInDiv({innerText: 'Open Sleeping Tabs'}), keys));
 		if (c.name === '_execute_browser_action') bindings.append(wrapInDiv('flex', wrapInDiv({innerText: 'Open Popup'}), keys));
 	});
 	if (document.getElementById('shortcut').classList.contains('show')) {